backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
pydantic==2.7.1
pydantic-settings==2.2.1
aiosqlite==0.20.0
datasets==2.19.1
huggingface_hub==0.23.0
cachetools==5.3.3
python-dotenv==1.0.1
httpx==0.27.0
soundfile==0.12.1
librosa==0.10.2
